+++ WP-PHP7.4/woo-scalexpert/Controller/Front/ProductController.php
@@ -60,8 +60,6 @@
             } else {
                 $solutions = $this->getEligibleSolutions( $price, $productID, '' );
             }
-            $groupFinancingSolution = get_option( "sg_scalexpert_group_financing_solution" );
-            
             
             echo '<!-- begin /Views/' . $template . '.php -->';
             if ( !$solutions && $template == 'payment-buttons' ) {
@@ -84,19 +82,6 @@
                         include( plugin_dir_path( __FILE__ ) . '../../Views/' . $template . '.php' );
                     }
                 }
-            }
-            // Paiements groupés / non groupés
-            if ( $groupFinancingSolution[ 'group_financing_solution' ] != 1 ) {
-                ?>
-                <script>
-                    jQuery(document).ready(function () {
-                        jQuery("div.payment_method_scalexpert").attr("style", "block");
-                    });
-                    jQuery("input[name=solutionCode]:radio").click(function () {
-                        jQuery("#payment_method_scalexpert").attr('checked', true);
-                    });
-                </script>
-                <?php
             }
             echo '<!-- end ' . plugin_dir_path( __FILE__ ) . '/Views/' . $template . '.php -->';
         }